VAE_ADHD/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.2.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2022.6.15=py37h06a4308_0
cffi=1.15.0=py37h7f8727e_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=37.0.1=py37h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
entrypoints=0.4=py37h06a4308_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.37.1=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.1=pypi_0
fsspec=2022.8.2=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.11.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.48.1=pypi_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py37h06a4308_0
ipython=7.31.1=py37h06a4308_1
jedi=0.18.1=py37h06a4308_1
joblib=1.1.0=pypi_0
jpeg=9e=h7f8727e_0
jupyter_client=7.2.2=py37h06a4308_0
jupyter_core=4.10.0=py37h06a4308_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lxml=4.9.1=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
monai=0.9.1=pypi_0
multidict=6.0.2=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py37h06a4308_0
nettle=3.7.3=hbbd107a_1
nibabel=4.0.1=pypi_0
nilearn=0.9.2=pypi_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
oauthlib=3.2.0=pypi_0
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.3.5=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py37hace64e9_1
pip=21.2.2=py37h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.4=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py37_1
python=3.7.6=h0371630_2
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.11.0=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.7.4=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.2.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py37h6a678d5_0
readline=7.0=h7b6447c_5
requests=2.28.1=py37h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.11.2=pypi_0
setuptools=61.2.0=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=h62c20be_0
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.11.0=py37_cu113
torchmetrics=0.9.3=pypi_0
torchvision=0.12.0=py37_cu113
tornado=6.1=py37h27cfd23_0
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=4.3.0=py37h06a4308_0
urllib3=1.26.11=py37h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
yarl=1.8.1=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.8.1=pypi_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
